[frontend] Rename auth components
