+++ frontend/app/(guest-only)/login/2fa/page.tsx
@@ -1,4 +1,4 @@
-import TwoFactorAuthVerifyForm from "@/app/settings/2fa/verify-form";
+import TwoFactorAuthVerifyForm from "@/app/ui/auth/verify-form";
 
 export default function TwoFactorAuthenticationPage() {
   return (
