+++ frontend/app/(guest-only)/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginForm from "@/app/ui/login-form";
+import LoginForm from "@/app/ui/auth/login-form";
 
 export default function LoginPage() {
   return <LoginForm />;
