+++ frontend/app/(guest-only)/signup/page.tsx
@@ -1,5 +1,5 @@
 // components
-import SignUpForm from "@/app/ui/signup-form";
+import SignUpForm from "@/app/ui/auth/signup-form";
 
 export default function SignUp() {
   return <SignUpForm />;
