+++ frontend/app/ui/auth/login-form.tsx
@@ -6,7 +6,7 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { useFormState, useFormStatus } from "react-dom";
-import { useAuthContext } from "../lib/client-auth";
+import { useAuthContext } from "@/app/lib/client-auth";
 
 export default function LoginForm() {
   return (
